+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
  - Use `$STEAM_COMPAT_DATA_PATH/pfx` as the game's prefix path if `STEAM_COMPAT_DATA_PATH` environment variable is set
 
+### Changed
+ - Winetricks is no longer required when running non-Winetricks commands (eg. `--command` or `--list`)
+
 ### Fixed
  - Fix detection for custom compatibility tools installed outside Steam directory
  - Fix detection for custom compatibility tool placing their manifest directly under `compatibilitytools.d` directory